quan ly nap tien

+++ database/migrations/2025_05_17_172203_create_deposit_history_table.php
@@ -16,14 +16,14 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('ma_nguoi_dung');
             $table->decimal('so_tien',15,2);
-            $table->decimal('khuyen_mai',15,2)->nullable();
-            $table->decimal('thuc_nhan',15,2);
+            $table->float('khuyen_mai')->nullable();
+            $table->float('thuc_nhan');
             $table->string('phuong_thuc');
             $table->string('ma_giao_dich')->unique();
             $table->string('trang_thai');
             $table->text('ghi_chu')->nullable();
             $table->timestamp('ngay_nap')->useCurrent();
-             $table->foreign('ma_nguoi_dung')->references('MaNguoiDung')->on('users')->onDelete('cascade');
+            $table->foreign('ma_nguoi_dung')->references('MaNguoiDung')->on('users')->onDelete('cascade');
         });
     }
